UPSTREAM: 4516: Infoblox - fix creation of CNAME/TXT records with empty view
This commit addresses an issue introduced in PR#3560.

**Problem Description:**
The search parameters are now always set for requests to WAPI.
This includes the view parameter, even if it's an empty string (the default value).
CNAME and TXT record types embed the view as a "string pointer" type,
so even an empty string is included in the request body to the WAPI.

The WAPI treats a view with an empty string value differently from a view that is not set at all.
The latter case results in using the default view, while the former tries to get a view
with an empty string value, causing an error: "View  not found".

**Fix:**
The fix is to not add the view pointer to the object that will be marshalled into the request body.
This will effectively skip the view parameter and allow the WAPI to use the default view.