NE-2053: Add dedicated Containerfile for Konflux builds
This commit adds a dedicated Containerfile that can be referenced
by the Konflux operand component. It automatically detects drift
from the `Dockerfile.openshift`, thereby enforcing synchronization between
the version used in CI presubmits and Konflux.

The base image used matches the minor version that was in use in CPaaS
at the time the migration to Konflux began. This ensures continuity with previous releases.

The builder image, on the other hand, has been set to `registry.access.redhat.com/ubi9/go-toolset`
to avoid relying on `registry-proxy.engineering.redhat.com` (used in CPaaS) and
to stay consistent with the builder image samples recommended by the Konflux migration team.